cambio de varibale

+++ taller_backend/DTOs/DetalleTipoPersonaDTOscs.cs
@@ -1,5 +1,6 @@
 ﻿namespace taller_backend.DTOs
 {
-    public record DetalleTipoPersonaCreateDto(string? Descripcion, int? IdTipoPersona);
-    public record DetalleTipoPersonaUpdateDto(string? Descripcion, int? IdTipoPersona);
+    public record DetalleTipoPersonaCreateDto(string? Descripcion, long? IdTipoPersona);
+    public record DetalleTipoPersonaUpdateDto(string? Descripcion, long? IdTipoPersona);
+
 }
